requirements.txt
torch
torchvision
scikit-learn
streamlit
numpy
matplotlib
joblib
Pillow
